***IMPORTANT*** change Profile Position Mode so that DS402 devices move to the next position setpoint immediately, without waiting to stop at the current one.
The reason we made this change is because we are sending our linear actuator a string of positions, all in the same direction. The actuator moves in a choppy way, because it quickly accelerates to the next setpoint, then decelerates and stops, before moving on.

This change is intended to make the actuator move smoothly between a string of setpoints that are in the same direction.

***It will also have the effect of making any DS402 devices immediately change direction without stopping, if the device is given a new setpoint which would take it in the opposite direction of where it is currently moving.***