Add option for 402 "Change on Setpoint".
This option is intended to be used on 402 devices in Profile Position mode. If this option is enabled, it's expected that if the device is currently moving to a setpoint, and then receives a new one telling it to move further in the same direction, then it will continue on to the new setpoint without stopping at the current one.

Reference for "Change on Setpoint":  page 11 of https://www.quicksilvercontrols.com/SP/AN/QCI-AN060_CiA402-CANopenDrivesAndMotionControlProfile.pdf